Create a Home component for Videossliders and manage the react routes
